Add like and dislike functionality
Notification popup at the bottom of the page after submitting answer on the question page.
New 'rating' field added to Account collection. 'rating' is an object with utorid as the key and 0 or 1 as the value.
New route added to update rating field.
A checkmark is rendered beside the question name in the question board page if a certain like/dislike ratio is met.
Additional files have been updated to include new field.